chore(fed): setup and admin dirs
