+++ apps/guardian-ui/src/setup/FederationSetup.tsx
@@ -8,17 +8,17 @@ import {
   VStack,
   Spinner,
 } from '@chakra-ui/react';
-import { ReactComponent as ArrowLeftIcon } from './assets/svgs/arrow-left.svg';
-import { Header } from './components/Header';
-import { useSetupContext } from './hooks';
+import { ReactComponent as ArrowLeftIcon } from '../assets/svgs/arrow-left.svg';
+import { Header } from '../components/Header';
+import { useSetupContext } from '../hooks';
 import { GuardianRole, SetupProgress, SETUP_ACTION_TYPE } from './types';
-import { RoleSelector } from './components/RoleSelector';
-import { SetConfiguration } from './components/SetConfiguration';
-import { Login } from './components/Login';
-import { ConnectGuardians } from './components/ConnectGuardians';
-import { RunDKG } from './components/RunDKG';
-import { VerifyGuardians } from './components/VerifyGuardians';
-import { SetupComplete } from './components/SetupComplete';
+import { RoleSelector } from '../components/RoleSelector';
+import { SetConfiguration } from '../components/SetConfiguration';
+import { Login } from '../components/Login';
+import { ConnectGuardians } from '../components/ConnectGuardians';
+import { RunDKG } from '../components/RunDKG';
+import { VerifyGuardians } from '../components/VerifyGuardians';
+import { SetupComplete } from '../components/SetupComplete';
 
 const PROGRESS_ORDER: SetupProgress[] = [
   SetupProgress.Start,
